[REFACTOR] Consolidate VerificationReasoner into reasoner.py
All Reasoner classes now live in a single module, matching the
A3BLReasoner layout. Sections are separated by double-line ===
banners so the file remains scannable.

- Move enumerate_label_assignments, top_k_satisfying and
  VerificationReasoner from ablkit/reasoning/verification.py into
  ablkit/reasoning/reasoner.py; delete the standalone file.
- Update ablkit/reasoning/__init__.py and ablkit/bridge/verification_bridge.py
  to import from .reasoner.
- Update docs/API/ablkit.reasoning.rst autofunction paths.
