+++ ablkit/reasoning/reasoner.py
@@ -5,8 +5,9 @@ between the knowledge base and learning models.
 Copyright (c) 2024 LAMDA.  All rights reserved.
 """
 
+import heapq
 import inspect
-from typing import Any, Callable, List, Optional, Union, Tuple
+from typing import Any, Callable, Iterator, List, Optional, Tuple, Union
 
 import numpy as np
 from zoopt import Dimension, Objective, Opt, Parameter, Solution
@@ -399,6 +400,11 @@ class Reasoner:
         return self.batch_abduce(data_examples)
 
 
+# =============================================================================
+# A3BL: Ambiguity-Aware Abductive Learning
+# =============================================================================
+
+
 class A3BLReasoner(Reasoner):
     """
     Reasoner for minimizing the inconsistency between the knowledge base and learning models.
@@ -601,3 +607,215 @@ class A3BLReasoner(Reasoner):
 
     def __call__(self, data_examples: ListData) -> List[List[Any]]:
         return self.batch_abduce(data_examples)
+
+
+# =============================================================================
+# Verification Learning
+#
+# Walks the per-symbol probability lattice in descending joint-probability
+# order, collecting the first top_k assignments that satisfy the knowledge
+# base. Reference: https://github.com/VerificationLearning/VerificationLearning
+# =============================================================================
+
+
+def enumerate_label_assignments(
+    pred_prob: np.ndarray, max_iter: int = 10000
+) -> Iterator[Tuple[List[int], float, List[float]]]:
+    """
+    Yield label-index assignments for a single data example in descending
+    joint-probability order. The walk is a Lawler-style best-first search:
+    each state is the tuple of per-symbol rank indices, and successors are
+    generated by advancing any one symbol to its next-best class.
+
+    Parameters
+    ----------
+    pred_prob : np.ndarray
+        Per-symbol probability matrix with shape ``(num_symbols, num_classes)``.
+    max_iter : int, optional
+        Hard cap on the number of yields. Defaults to 10000.
+
+    Yields
+    ------
+    labels : List[int]
+        Class indices for each symbol.
+    joint_prob : float
+        Product of the chosen per-symbol probabilities.
+    per_symbol_probs : List[float]
+        The chosen probability for each symbol.
+    """
+    pred_prob = np.asarray(pred_prob, dtype=float)
+    num_symbols, num_classes = pred_prob.shape
+    if num_symbols == 0:
+        return
+
+    sorted_indices = np.argsort(-pred_prob, axis=1)
+    sorted_probs = np.take_along_axis(pred_prob, sorted_indices, axis=1)
+
+    initial_state = (0,) * num_symbols
+    initial_prob = float(np.prod(sorted_probs[:, 0]))
+
+    seen = {initial_state}
+    heap: List[Tuple[float, Tuple[int, ...]]] = [(-initial_prob, initial_state)]
+    yields = 0
+
+    while heap and yields < max_iter:
+        neg_prob, state = heapq.heappop(heap)
+        joint_prob = -neg_prob
+        labels = [int(sorted_indices[i, state[i]]) for i in range(num_symbols)]
+        per_symbol_probs = [float(sorted_probs[i, state[i]]) for i in range(num_symbols)]
+        yield labels, joint_prob, per_symbol_probs
+        yields += 1
+
+        for sym in range(num_symbols):
+            next_rank = state[sym] + 1
+            if next_rank >= num_classes:
+                continue
+            new_state = state[:sym] + (next_rank,) + state[sym + 1:]
+            if new_state in seen:
+                continue
+            seen.add(new_state)
+            current_p = sorted_probs[sym, state[sym]]
+            next_p = sorted_probs[sym, next_rank]
+            if current_p <= 0:
+                new_joint = 0.0
+            else:
+                new_joint = joint_prob * (next_p / current_p)
+            heapq.heappush(heap, (-new_joint, new_state))
+
+
+def top_k_satisfying(
+    pred_prob: np.ndarray,
+    predicate: Callable[[List[Any]], bool],
+    top_k: int = 1,
+    max_iter: int = 10000,
+    idx_to_label: Optional[dict] = None,
+) -> Tuple[List[List[Any]], List[float]]:
+    """
+    Walk label assignments in descending joint-probability order and return
+    the first ``top_k`` that satisfy ``predicate``. If none is found within
+    ``max_iter`` iterations the single highest-probability assignment is
+    returned as a fallback so callers always receive a usable label.
+
+    Parameters
+    ----------
+    pred_prob : np.ndarray
+        Per-symbol probability matrix with shape ``(num_symbols, num_classes)``.
+    predicate : Callable[[List[Any]], bool]
+        Function called on each candidate label sequence; truthy means the
+        candidate is consistent with the knowledge base.
+    top_k : int, optional
+        Maximum number of satisfying candidates to return. Defaults to 1.
+    max_iter : int, optional
+        Hard cap on enumeration steps. Defaults to 10000.
+    idx_to_label : dict, optional
+        Optional mapping from class index to pseudo-label. When omitted, the
+        raw class indices are returned.
+
+    Returns
+    -------
+    candidates : List[List[Any]]
+        Label assignments that satisfy ``predicate`` (or the fallback).
+    probs : List[float]
+        Joint probability of each returned candidate.
+    """
+    matches: List[List[Any]] = []
+    probs: List[float] = []
+    fallback: Optional[Tuple[List[Any], float]] = None
+
+    for labels_idx, joint_prob, _ in enumerate_label_assignments(pred_prob, max_iter):
+        labels = (
+            [idx_to_label[i] for i in labels_idx] if idx_to_label is not None else labels_idx
+        )
+        if fallback is None:
+            fallback = (labels, joint_prob)
+        if predicate(labels):
+            matches.append(labels)
+            probs.append(joint_prob)
+            if len(matches) >= top_k:
+                break
+
+    if not matches and fallback is not None:
+        matches.append(fallback[0])
+        probs.append(fallback[1])
+    return matches, probs
+
+
+class VerificationReasoner:
+    """
+    Reasoner used by :class:`~ablkit.bridge.VerificationBridge`. Rather than
+    picking a single best candidate via a distance function, it enumerates
+    the top ``top_k`` label assignments that satisfy the knowledge base,
+    ordered by joint probability. The bridge then trains the model on each
+    of those candidates.
+
+    Parameters
+    ----------
+    kb : KBBase
+        The knowledge base used to verify candidates. ``kb.logic_forward``
+        must return the reasoning result so it can be compared with each
+        data example's ``Y``.
+    top_k : int, optional
+        Number of satisfying candidates to enumerate per example.
+        Defaults to 1.
+    max_iter : int, optional
+        Maximum number of enumeration steps per example before giving up
+        and returning the fallback. Defaults to 10000.
+    idx_to_label : dict, optional
+        A mapping from base-model index to pseudo-label. If omitted a
+        default order-based mapping is built from ``kb.pseudo_label_list``.
+    """
+
+    def __init__(
+        self,
+        kb: KBBase,
+        top_k: int = 1,
+        max_iter: int = 10000,
+        idx_to_label: Optional[dict] = None,
+    ) -> None:
+        if top_k < 1:
+            raise ValueError("top_k must be >= 1.")
+        if max_iter < 1:
+            raise ValueError("max_iter must be >= 1.")
+        self.kb = kb
+        self.top_k = top_k
+        self.max_iter = max_iter
+        if idx_to_label is None:
+            idx_to_label = dict(enumerate(kb.pseudo_label_list))
+        self.idx_to_label = idx_to_label
+        self.label_to_idx = {label: idx for idx, label in idx_to_label.items()}
+
+    def top_k_candidates(
+        self, pred_prob: np.ndarray, y: Any
+    ) -> Tuple[List[List[Any]], List[float]]:
+        """
+        Return up to ``top_k`` label assignments for one data example whose
+        ``kb.logic_forward`` matches ``y``.
+        """
+
+        def predicate(labels: List[Any]) -> bool:
+            return self.kb.logic_forward(labels) == y
+
+        return top_k_satisfying(
+            pred_prob,
+            predicate,
+            top_k=self.top_k,
+            max_iter=self.max_iter,
+            idx_to_label=self.idx_to_label,
+        )
+
+    def batch_top_k(self, data_examples) -> List[List[List[Any]]]:
+        """
+        Run :meth:`top_k_candidates` on every example in ``data_examples``.
+        Stores the result on ``data_examples.top_k_candidates`` and
+        ``data_examples.top_k_probs``. Returns the list of per-example
+        candidate lists.
+        """
+        all_candidates: List[List[List[Any]]] = []
+        all_probs: List[List[float]] = []
+        for data_example in data_examples:
+            cands, probs = self.top_k_candidates(data_example.pred_prob, data_example.Y)
+            all_candidates.append(cands)
+            all_probs.append(probs)
+        data_examples.top_k_candidates = all_candidates
+        data_examples.top_k_probs = all_probs
+        return all_candidates
